Crate timestamp

Allow block authors to include a timestamp via an inherent transaction.

This is roughly analogous to FRAME’s pallet timestamp. It relies on the same client-side inherent data provider, as well as Tuxedo’s own previous block inehrent data provider.

In each block, the block author must include a single SetTimestamp transaction that peeks at the Timestamp UTXO that was created in the previous block, and creates a new one with an updated timestamp.

This piece currently features two prominent hacks which will need to be cleaned up in due course.

It abuses the UpForGrabs verifier. This should be replaced with an Unspendable verifier and an eviction workflow.

In block #1 it allows creating a new best timestamp without comsuming a previous one. This should be removed once we are able to include a timestamp in the genesis block.

Structs

CleanUpTimestamp

Allows users to voluntarily clean up old timestamps by showing that there exists another timestamp that is at least the CLEANUP_AGE newer.

SetTimestamp

A constraint checker for the simple act of setting a new best timetamp.

Timestamp

A timestamp, since the unix epoch, noted at some point in the history of the chain. It also records the block height in which it was included.

Enums

TimestampError

Reasons that setting or cleaning up the timestamp may go wrong.

Traits

TimestampConfig

Options to configure the timestamp piece in your runtime. Currently we only need access to a block number.